Disable selecting competition if result is loading
